requirements.txt
django
mysqlclient
djangorestframework
djoser